+++ packages/ciphernode/enclave/Cargo.toml
@@ -4,36 +4,22 @@ version = "0.1.0"
 edition = "2021"
 description = ": coordinates the encryption and decryption of enclave computations"
 repository = "https://github.com/gnosisguild/enclave/packages/ciphernode"
-build = "build.rs"
 
 [dependencies]
 actix = { workspace = true }
-alloy = { workspace = true }
 anyhow = { workspace = true }
 cipher = { path = "../cipher" }
 clap = { workspace = true }
 compile-time = { workspace = true }
 config = { path = "../config" }
-data = { path = "../data" }
 dialoguer = "0.11.0"
-dirs = { workspace = true }
 enclave-core = { path = "../core" }
 enclave_node = { workspace = true }
-evm = { workspace = true }
 hex = { workspace = true }
-libp2p = { workspace = true }
-net = { workspace = true }
-once_cell = "1.20.2"
 petname = "2.0.2"
-phf = { version = "0.11", features = ["macros"] }
 rand = { workspace = true }
-e3_request = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true } 
 zeroize = { workspace = true }
 
-[build-dependencies]
-serde_json = { workspace = true }
