+++ packages/ciphernode/enclave/src/net_purge.rs
@@ -1,14 +1,9 @@
-use actix::Actor;
 use anyhow::*;
 use config::AppConfig;
-use enclave_core::EventBus;
-use enclave_node::get_repositories;
-use net::NetRepositoryFactory;
+use enclave_node::net_purge;
 
 pub async fn execute(config: &AppConfig) -> Result<()> {
-    let bus = EventBus::new(true).start();
-    let repositories = get_repositories(&config, &bus)?;
-    repositories.libp2p_keypair().clear();
+    net_purge::execute(config).await?;
     println!("Peer ID has been purged. A new Peer ID will be generated upon restart.");
     Ok(())
 }
