+++ packages/ciphernode/enclave/src/start.rs
@@ -2,7 +2,7 @@ use crate::owo;
 use anyhow::{anyhow, Result};
 use config::AppConfig;
 use enclave_core::get_tag;
-use enclave_node::{listen_for_shutdown, setup_ciphernode};
+use enclave_node::{listen_for_shutdown, start};
 use tracing::{info, instrument};
 
 #[instrument(name="app", skip_all,fields(id = get_tag()))]
@@ -12,7 +12,7 @@ pub async fn execute(config: AppConfig) -> Result<()> {
         return Err(anyhow!("You must provide an address"));
     };
 
-    let (bus, handle, peer_id) = setup_ciphernode(config, address).await?;
+    let (bus, handle, peer_id) = start::execute(config, address).await?;
     info!("LAUNCHING CIPHERNODE: ({}/{})", address, peer_id);
 
     tokio::spawn(listen_for_shutdown(bus.into(), handle));
