+++ packages/ciphernode/enclave_node/src/lib.rs
@@ -1,9 +1,11 @@
-mod aggregator;
-mod ciphernode;
-mod datastore;
-mod shutdown;
+pub mod aggregator_start;
+mod helpers;
+pub mod init;
+pub mod net_generate;
+pub mod net_purge;
+pub mod net_set;
+pub mod password_create;
+pub mod start;
+pub mod wallet_set;
 
-pub use aggregator::*;
-pub use ciphernode::*;
-pub use datastore::*;
-pub use shutdown::*;
+pub use helpers::*;
