+++ packages/ciphernode/enclave_node/src/net_purge.rs
@@ -1,0 +1,13 @@
+use crate::datastore::get_repositories;
+use actix::Actor;
+use anyhow::*;
+use config::AppConfig;
+use enclave_core::EventBus;
+use net::NetRepositoryFactory;
+
+pub async fn execute(config: &AppConfig) -> Result<()> {
+    let bus = EventBus::new(true).start();
+    let repositories = get_repositories(&config, &bus)?;
+    repositories.libp2p_keypair().clear();
+    Ok(())
+}
